refactor(crud): consolidate templates and update generation paths
- Move all CRUD templates to unified crud directory
- Update template paths in controllers, services, and validations generators
- Reorganize generated file locations with common/ prefix for controllers and services
- Add users service directory to project structure
- Sort package.json scripts alphabetically for better maintainability
- Fix test paths to reflect new template organization
