+++ context/structure.txt
@@ -170,6 +170,7 @@
 │   ├── bot
 │   ├── common
 │   ├── desktop
+│   ├── users
 │   ├── wearable
 │   └── web
 ├── templates
@@ -248,4 +249,4 @@
     ├── extensions.json
     └── settings.json
 
-89 directories, 160 files
+90 directories, 160 files
