+++ package.json
@@ -31,12 +31,12 @@
   "type": "commonjs",
   "main": "index.js",
   "bin": {
+    "generate-controllers": "./scripts/generateControllers.script.js",
     "generate-crud-docs": "./scripts/generateCrudDocs.script.js",
     "generate-crud-endpoints": "./scripts/generateCrudEndpoints.script.js",
-    "generate-models": "./scripts/createModels.script.js",
     "generate-crud-validations": "./scripts/generateCrudValidations.script.js",
-    "generate-services": "./scripts/generateServices.script.js",
-    "generate-controllers": "./scripts/generateControllers.script.js"
+    "generate-models": "./scripts/createModels.script.js",
+    "generate-services": "./scripts/generateServices.script.js"
   },
   "config": {
     "commitizen": {
@@ -47,24 +47,24 @@
     "**/*": ["biome check --fix --unsafe", "biome format --write", "git add ."]
   },
   "scripts": {
+    "debug": "node ./scripts/debug.script.js",
     "dev": "node --watch index.js",
     "dev:inspect": "node --watch --inspect index.js",
-    "debug": "node ./scripts/debug.script.js",
     "format": "biome format",
     "format:write": "biome format --write",
     "lint": "biome lint",
     "lint:fix": "biome lint --fix",
+    "models:create": "node ./scripts/createModels.script.js",
+    "models:generate": "node ./scripts/createModels.script.js",
     "prepare": "husky",
     "start": "node index.js",
+    "sync-endpoints": "node ./scripts/endpoints.script.js",
     "test": "NODE_ENV=test jest --detectOpenHandles --passWithNoTests",
-    "test:unit": "jest tests/unit --passWithNoTests",
-    "test:watch": "jest --watch",
-    "test:coverage": "NODE_ENV=test jest --coverage",
-    "test:verbose": "jest --verbose",
     "test:ci": "jest --ci --reporters=default --reporters=jest-junit",
-    "models:generate": "node ./scripts/createModels.script.js",
-    "models:create": "node ./scripts/createModels.script.js",
-    "sync-endpoints": "node ./scripts/endpoints.script.js"
+    "test:coverage": "NODE_ENV=test jest --coverage",
+    "test:unit": "jest tests/unit --passWithNoTests",
+    "test:verbose": "jest --verbose",
+    "test:watch": "jest --watch"
   },
   "overrides": {
     "glob": "^10.3.10",
