+++ scripts/generateServices.script.js
@@ -146,7 +146,7 @@ class CrudServicesGenerator {
 
   async generateService(tableData, singularName, pluralName) {
     try {
-      let serviceContent = await this.crudHelper.getTemplate('services', 'services');
+      let serviceContent = await this.crudHelper.getTemplate('crud', 'services');
 
       const mainModelName = toCamelCase(tableData.tableName);
       const serviceName = `${formatCapitalize(singularName)}Services`;
@@ -274,7 +274,7 @@ class CrudServicesGenerator {
   async saveService(serviceContent, groupName, pluralName) {
     try {
       const fileName = `${toCamelCase(pluralName)}.services`;
-      const folderPath = await this.crudHelper.createFolder('SERVICES', groupName, '');
+      const folderPath = await this.crudHelper.createFolder('SERVICES', 'common/' + groupName, '');
       const filePath = await this.crudHelper.createFile(folderPath, fileName, serviceContent);
 
       console.log(`📄 Service saved to: ${filePath}`);
