+++ templates/crud/services.template.js
@@ -7,9 +7,9 @@ const { Op } = require('sequelize');
 // INTERNAL DEPENDENCIES
 // =============================================================================
 const LogServices = require('../logs/logs.service');
-const sequelize = require('../../config/database/connection');
-const { bulkToggleSoftDelete, paginateModel, setSearchQuery } = require('../../helpers/database.helper');
-const { wrapLogging } = require('../../helpers/debug.helper');
+const sequelize = require('../../../config/database/connection');
+const { bulkToggleSoftDelete, paginateModel, setSearchQuery } = require('../../../helpers/database.helper');
+const { wrapLogging } = require('../../../helpers/debug.helper');
 
 // =============================================================================
 // MODELS
