+++ templates/crud/validations.template.js
@@ -1,6 +1,6 @@
-const { {{MAIN_MODEL}}, {{MORE_MODELS}} } = require('../../../config/database/connection').models;
-const { commonSchemas, databaseSchemas } = require('../../../helpers/validations');
-const { paginationSchemas, searchSchemas, filtersSchemas } = require('../../../schemas/validations');
+const { {{MAIN_MODEL}}, {{MORE_MODELS}} } = require('../../../../config/database/connection').models;
+const { commonSchemas, databaseSchemas } = require('../../../../helpers/validations');
+const { paginationSchemas, searchSchemas, filtersSchemas } = require('../../../../schemas/validations');
 
 const createSchema = {
   // Add any additional body parameters here
